requirements.txt
pyusb==1.0.0
flask>=0.10.1
